requirements.txt
flask==2.3.3
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
click==8.1.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0